+++ ui/src/components/dash-chart-widget/index.js
@@ -1,0 +1,1 @@
+export { default } from "./DashChartWidget.vue";
