+++ ui/src/components/dash-chart-widget/widgets/index.js
@@ -1,0 +1,1 @@
+export { RecordTypes } from "./record-types";
