--- ui/src/components/dash-chart/index.js
@@ -1,1 +1,0 @@
-export { default } from "./DashChart.vue";
